Изменение типа возвращаеммой переменной

+++ Drivers/BSP/Inc/stm32f3_discovery_accelerometer.h
@@ -91,7 +91,7 @@ ACCELERO_StatusTypeDef;
 /* Acc functions */  
 uint8_t   BSP_ACCELERO_Init(void);
 void      BSP_ACCELERO_Reset(void);
-void      BSP_ACCELERO_GetXYZ(int16_t *pDataXYZ);
+void      BSP_ACCELERO_GetXYZ(float *pDataXYZ);
 
 /**
   * @}
